Remove spaces from posts app
