+++ SWE574_project/urls.py
@@ -22,6 +22,7 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("", include("posts.urls")),
     path("", include("users.urls")),
+    path("", include("spaces.urls")),
     path("", include("django.contrib.auth.urls")),
 ]
 if settings.DEBUG:
