+++ posts/admin.py
@@ -1,10 +1,6 @@
 from django.contrib import admin
 
-from users.models import Contact
-
-from .models import Comment, Post, Space
-
-admin.site.register(Post)
+from .models import Comment, Post
 
 
 class PostAdmin(admin.ModelAdmin):
@@ -25,20 +21,11 @@ class PostAdmin(admin.ModelAdmin):
     ordering = ["status", "publish"]
 
 
-admin.site.register(Comment)
-admin.site.register(Space)
-
-
 class CommentAdmin(admin.ModelAdmin):
     list_display = ("author", "post", "created_date", "approved_comment")
     list_filter = ("created_date", "post")
     search_fields = ("author", "post", "text")
 
 
-# admin.site.register(Author)
-#
-#
-# class AuthorAdmin(admin.ModelAdmin):
-#     list_display = ['user', 'date_of_birth', 'photo']
-
-admin.site.register(Contact)
+admin.site.register(Post)
+admin.site.register(Comment)
