+++ posts/forms.py
@@ -1,35 +1,16 @@
 from django import forms
 from django.contrib.auth import get_user_model
 
-from posts.models import Comment, Post, Space
+from posts.models import Comment, Post
 from users.models import Profile
 
 User = get_user_model()
-
-
-class SpaceCreationForm(forms.ModelForm):
-    class Meta:
-        model = Space
-        fields = [
-            "name",
-            "description",
-            "members",
-            "moderators",
-            "is_all_members_post_allowed",
-            "is_only_moderators_post_allowed",
-        ]
 
 
 class PostForm(forms.ModelForm):
     class Meta:
         model = Post
         fields = ("title", "link", "tags", "labels", "upload", "text", "image")
-
-
-class SpaceForm(forms.ModelForm):
-    class Meta:
-        model = Post
-        fields = ("title", "link", "text")
 
 
 class CommentForm(forms.ModelForm):
