+++ spaces/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from spaces.models import  Space
+
+
+admin.site.register(Space)
